+++ packages/lumx-vue/src/components/flag/index.ts
@@ -1,0 +1,8 @@
+import { CLASSNAME, FlagProps } from '@lumx/core/js/components/Flag';
+
+import Flag from './Flag.vue';
+
+Flag.className = CLASSNAME;
+
+export { Flag };
+export type { FlagProps };